refactor: examples: rename network module to http_api
# Summary

In raft-kv-memstore the `network` module only held the application's HTTP read
handlers, while the actual Raft network lives in the `network-v2-http` crate, so
the name misled readers. Rename it to `http_api` and collapse the now-redundant
single-child module into one file.

# Details

- Move `src/network/api.rs` to `src/http_api.rs` and drop the `network/mod.rs`
  wrapper that existed only to re-export `api`.
- Point the `lib.rs` module declaration and the `/read`, `/linearizable_read`,
  and `/follower_read` route registrations at `http_api`.
- Update the README code tour to the new path.
