+++ examples/raft-kv-memstore/src/http_api.rs
